requirements.txt
Django>=5,<6
gunicorn
psycopg[binary]>=3.1
whitenoise 
dj-database-url
python-dotenv==1.1.1
djangorestframework>=3.14
